fix: 初步完成工作台页面重构
