+++ package.json
@@ -19,7 +19,6 @@
     "@hi-ui/core-css": "^2.0.0",
     "@hi-ui/hiui": "^3.0.0",
     "axios": "^0.18.1",
-    "axios-mock-adapter": "^1.19.0",
     "classnames": "^2.2.5",
     "echarts": "^4.2.1",
     "echarts-for-react": "^2.0.15-beta.0",
