+++ src/template/home-workbench/oldIndex.scss
@@ -1,0 +1,75 @@
+@import '@hi-ui/core-css/index.scss';
+
+.page--workbench {
+  padding: $spacer-4 $spacer-6;
+  min-height: calc(100vh - 64px);
+  box-sizing: border-box;
+
+  .workbench--container {
+    max-width: 1300px;
+    margin: 0 auto;
+  }
+
+  .user {
+    line-height: 36px;
+    color: $black;
+
+    &__icon {
+      display: inline-block;
+      width: 36px;
+      height: 36px;
+      margin-right: $spacer-3;
+      background: $white;
+      border: 1px solid $gray-lighter;
+      border-radius: 50%;
+      text-align: center;
+    }
+  }
+
+  .hi-grid__col-8 .card {
+    height: 377px;
+
+    .card__body {
+      flex-basis: 377px;
+      padding: $spacer-3 $spacer-6 $spacer-6 $spacer-6;
+    }
+  }
+
+  .hi-grid__col-16 .card {
+    height: 377px;
+
+    .card__body {
+      flex-basis: 377px;
+
+      .hi-grid__row {
+        width: 100%;
+      }
+
+      .hi-timeline {
+        margin-top: 14px;
+        width: 100%;
+      }
+    }
+  }
+
+  .todo {
+    width: 100%;
+    color: $black;
+
+    &__item {
+      padding: 9px 0;
+
+      .approval-btn {
+        color: get-color($palette-hiui, '50');
+      }
+    }
+  }
+
+  .hi-dropdown__title-text {
+    padding-right: $spacer-1;
+  }
+
+  .hi-stepper {
+    padding: $spacer-6 0 $spacer-5 0;
+  }
+}
